Move metadata parser to new file
